解决已批改作业百分数小数点后过长的问题,解决已批改作业预览界面少题的问题.

+++ edt/youngyedu/src/main/java/com/yougy/home/fragment/showFragment/ExerciseBookFragment.java
@@ -132,7 +132,7 @@ public class ExerciseBookFragment extends BFragment {
                         } else {
                             if (isMutualEvaluation(homeworkSummary)){
                                 holder.binding.statusTv.setVisibility(View.GONE);
-                                holder.binding.textRateScore.setText("正确率：" + (extraBean.getAccuracy()*100) + "%");
+                                holder.binding.textRateScore.setText("正确率：" + (int)(extraBean.getAccuracy()*100) + "%");
                             }
                             else {
                                 holder.binding.statusTv.setVisibility(View.GONE);
